requirements.txt
flask-restful
flask_cors
flask_restful_swagger
jsonschema
pyyaml
requests
pandas
sklearn
numpy
seaborn
nltk
plotly
chart_studio
cufflinks
